+++ src/routes/projects/+page.svelte
@@ -53,7 +53,7 @@
     />
     <AddProject
         add={(template) => {
-            const newProjectID = Projects.copy(template);
+            const newProjectID = Projects.copy(template, $user?.uid ?? null);
             goto(`/project/${newProjectID}`);
         }}
     />